Moved Main.js to be in the lib directory.
