+++ conway_module/lib/dom/DomUtilities.js
@@ -1,0 +1,23 @@
+/**
+ * Helper function for finding elements on the document.
+ * @param {*} id The ID of the element to fetch.
+ * Throws an error if the element could not be found.
+ */
+function getElementById(id) {
+	let element = document.getElementById(id);
+	if (!element) {
+		throw new Error(`No element could be found with the ID of ${id}`);
+	}
+	return element;
+}
+
+/**
+ * Helper function for setting an elements value.
+ * @param {string} id
+ * @param {*} value
+ */
+function setElementValue(id, value) {
+	getElementById(id).value = value;
+}
+
+module.exports = { getElementById, setElementValue };
